Ajustado a logica do projeto e criado CRUD de categoria e suas demais classes.

+++ src/main/java/desafioapirest/dio/Dtos/CategoriaDto.java
@@ -2,9 +2,8 @@ package desafioapirest.dio.Dtos;
 
 import desafioapirest.dio.domain.model.Categoria;
 
-import java.math.BigDecimal;
 
-public record CategoriaDto(Long id, String nome, BigDecimal limiteOrcamento) {
+public record CategoriaDto(Long id, String nome, Double limiteOrcamento) {
 
     public CategoriaDto(Categoria categoria) {
         this(categoria.getId(), categoria.getNome(), categoria.getLimiteOrcamento());
